made sorting mobile friendly
