+++ app/javascript/constants/GenreOptions.js
@@ -1,0 +1,36 @@
+const genreOptions = [
+  {
+    display: "Action",
+    value: "Action"
+  },
+  {
+    display: "Animation",
+    value: "Animation"
+  },
+  {
+    display: "Comedy",
+    value: "Comedy"
+  },
+  {
+    display: "Drama",
+    value: "Drama"
+  },
+  {
+    display: "Fantasy",
+    value: "Fantasy"
+  },
+  {
+    display: "Horror",
+    value: "Horror"
+  },
+  {
+    display: "Romance",
+    value: "Romance"
+  },
+  {
+    display: "Sci-Fi",
+    value: "Sci-Fi"
+  }
+]
+
+export default genreOptions
