+++ app/javascript/constants/RatingOptions.js
@@ -1,0 +1,24 @@
+const ratingOptions = [
+  {
+    display: "Five Stars",
+    value: 5
+  },
+  {
+    display: "Four Stars",
+    value: 4
+  },
+  {
+    display: "Three Stars",
+    value: 3
+  },
+  {
+    display: "Two Stars",
+    value: 2
+  },
+  {
+    display: "One Star",
+    value: 1
+  }
+]
+
+export default ratingOptions
