+++ app/javascript/constants/TypeOptions.js
@@ -1,0 +1,12 @@
+const typeOptions = [
+  {
+    display: "Movie",
+    value: "movie"
+  },
+  {
+    display: "Show",
+    value: "show"
+  }
+]
+
+export default typeOptions
